requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
sqlalchemy==2.0.23
starlette==0.27.0
python-dotenv==1.0.0
requests==2.31.0
backoff==2.2.1
tenacity==8.2.3
email-validator==2.1.0.post1